# packages/openapi/openapi.json
{"openapi":"3.0.0","paths":{"/users/{username}":{"get":{"operationId":"UsersController_findByUsername","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/posts":{"get":{"operationId":"PostsController_getPosts","parameters":[],"responses":{"200":{"description":""}},"tags":["Posts"]},"post":{"operationId":"PostsController_createPost","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDto"}}}},"responses":{"201":{"description":""}},"tags":["Posts"]}},"/posts/{slug}":{"get":{"operationId":"PostsController_getPostBySlug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Posts"]}},"/posts/{id}":{"get":{"operationId":"PostsController_getPostsByCommunityId","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Posts"]},"patch":{"operationId":"PostsController_updatePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostsDto"}}}},"responses":{"200":{"description":""}},"tags":["Posts"]},"delete":{"operationId":"PostsController_deletePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Posts"]}},"/posts/liked":{"get":{"operationId":"PostsController_getLikedPosts","parameters":[],"responses":{"200":{"description":""}},"tags":["Posts"]}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/signin":{"post":{"operationId":"AuthController_signIn","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/signin/google":{"get":{"operationId":"AuthController_signInWithGoogle","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/google/callback":{"get":{"operationId":"AuthController_googleCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/likes":{"post":{"operationId":"LikesController_unlikePost","parameters":[{"name":"postId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Likes"]}},"/communities":{"post":{"operationId":"CommunitiesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommunityDto"}}}},"responses":{"201":{"description":""}},"tags":["Communities"]},"get":{"operationId":"CommunitiesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Communities"]}},"/communities/{id}":{"get":{"operationId":"CommunitiesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Communities"]},"patch":{"operationId":"CommunitiesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCommunityDto"}}}},"responses":{"200":{"description":""}},"tags":["Communities"]},"delete":{"operationId":"CommunitiesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Communities"]}},"/comments":{"post":{"operationId":"CommentsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentDto"}}}},"responses":{"201":{"description":""}},"tags":["Comments"]},"get":{"operationId":"CommentsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Comments"]}},"/comments/post/{id}":{"get":{"operationId":"CommentsController_findByPostId","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Comments"]}},"/comments/{id}":{"get":{"operationId":"CommentsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Comments"]},"patch":{"operationId":"CommentsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCommentDto"}}}},"responses":{"200":{"description":""}},"tags":["Comments"]},"delete":{"operationId":"CommentsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Comments"]}}},"info":{"title":"Art Space","description":"This api is for the Art Space project","version":"1.0","contact":{}},"tags":[{"name":"nestjs","description":""}],"servers":[],"components":{"schemas":{"CreatePostDto":{"type":"object","properties":{}},"UpdatePostsDto":{"type":"object","properties":{}},"SignUpDto":{"type":"object","properties":{}},"CreateCommunityDto":{"type":"object","properties":{}},"UpdateCommunityDto":{"type":"object","properties":{}},"CreateCommentDto":{"type":"object","properties":{}},"UpdateCommentDto":{"type":"object","properties":{}}}}}